+++ Z.MVVMHelper/VmBase.cs
@@ -31,11 +31,12 @@ namespace Z.MVVMHelper
             }
         }
 
+        /// <inheritdoc />
         /// <summary>
         ///     Constructor for <see cref="VmBase" />
         /// </summary>
         protected VmBase() : this(true) { }
-
+        
         /// <summary>
         ///     Automate calls to <see cref="OnPropertyChanging" /> and <see cref="OnPropertyChanged" />
         /// </summary>
